Add files via upload
Shots on target vs Non-penalty goals minus Expected goals.  
The marks are labeled by player name.  
Players in this chart have atleast 10 games worth of minutes so far in the 22/23 season. 
Attackers that are inaccurate shot-takers and accurate finishers in Q1
Attackers that are accurate shot-takers and accurate finishers in Q2
Attackers that are inaccurate shot takers and inaccurate finishers in Q3
Attackers that are accurate shot takers and inaccurate finishers in Q4
Data was collected from FBref as of Thursday, March 9th, 2023 at 2:30 AM EST